+++ internal/broker/keygen.go
@@ -15,9 +15,13 @@
 package broker
 
 import (
+	"crypto/rand"
+	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"log"
+	"math"
+	"math/big"
 	"net/http"
 	"strconv"
 	"text/template"
@@ -25,6 +29,83 @@ import (
 
 	"github.com/emitter-io/emitter/internal/security"
 )
+
+// onKeyGen processes a keygen request.
+func (c *Conn) onKeyGen(payload []byte) (response, bool) {
+	// Deserialize the payload.
+	message := keyGenRequest{}
+	if err := json.Unmarshal(payload, &message); err != nil {
+		return ErrBadRequest, false
+	}
+
+	key, err := c.service.generateKey(message.Key, message.Channel, message.access(), message.expires())
+	if err != nil {
+		return err, false
+	}
+
+	// Success, return the response
+	return &keyGenResponse{
+		Status:  200,
+		Key:     key,
+		Channel: message.Channel,
+	}, true
+}
+
+func (s *Service) generateKey(rawMasterKey string, channel string, access uint8, expires time.Time) (string, *Error) {
+	// Attempt to parse the key, this should be a master key
+	masterKey, err := s.Cipher.DecryptKey([]byte(rawMasterKey))
+	if err != nil || !masterKey.IsMaster() || masterKey.IsExpired() {
+		return "", ErrUnauthorized
+	}
+
+	// Attempt to fetch the contract using the key. Underneath, it's cached.
+	contract, contractFound := s.contracts.Get(masterKey.Contract())
+	if !contractFound {
+		return "", ErrNotFound
+	}
+
+	// Validate the contract
+	if !contract.Validate(masterKey) {
+		return "", ErrUnauthorized
+	}
+
+	// Generate random salt
+	n, err := rand.Int(rand.Reader, big.NewInt(math.MaxInt16))
+	if err != nil {
+		return "", ErrServerError
+	}
+
+	// Create a key request
+	key := security.Key(make([]byte, 24))
+	key.SetSalt(uint16(n.Uint64()))
+	key.SetMaster(masterKey.Master())
+	key.SetContract(masterKey.Contract())
+	key.SetSignature(masterKey.Signature())
+	key.SetPermissions(access)
+	key.SetExpires(expires)
+
+	// Set the target and return an convert the error if it occurs
+	if err := key.SetTarget(channel); err != nil {
+		switch err {
+		case security.ErrTargetInvalid:
+			return "", ErrTargetInvalid
+		case security.ErrTargetTooLong:
+			return "", ErrTargetTooLong
+		default:
+			return "", ErrServerError
+		}
+	}
+
+	// Encrypt the final key
+	out, err := s.Cipher.EncryptKey(key)
+	if err != nil {
+		return "", ErrServerError
+	}
+
+	return out, nil
+}
+
+// ------------------------------------------------------------------------------------
 
 type keygenForm struct {
 	Key      string
@@ -46,9 +127,7 @@ func (f *keygenForm) parse(req *http.Request) bool {
 	var parsedOK = true
 
 	canParseToBool := func(source string) (bool, bool) {
-
 		value := req.FormValue(source)
-
 		if value == "" {
 			return false, true
 		} else if value == "on" {
@@ -61,9 +140,7 @@ func (f *keygenForm) parse(req *http.Request) bool {
 	}
 
 	canParseToInt := func(source string) (int64, bool) {
-
 		value := req.FormValue(source)
-
 		if value == "" {
 			return 0, true
 		}
@@ -111,7 +188,6 @@ func (f *keygenForm) parse(req *http.Request) bool {
 
 // validate required fields
 func (f *keygenForm) isValid() bool {
-
 	if f.Key == "" {
 		f.Response = "Missing SecretKey"
 		return false
@@ -135,7 +211,6 @@ func (f *keygenForm) expires() time.Time {
 
 func (f *keygenForm) access() uint8 {
 	required := security.AllowNone
-
 	if f.Sub {
 		required |= security.AllowRead
 	}
@@ -159,17 +234,17 @@ func (f *keygenForm) access() uint8 {
 }
 
 func handleKeyGen(s *Service) http.HandlerFunc {
-
 	var fs http.FileSystem = assets
-
 	f, err := fs.Open("keygen.html")
 	if err != nil {
 		panic(err)
 	}
+
 	b, err := ioutil.ReadAll(f)
 	if err != nil {
 		panic(err)
 	}
+
 	t, err := template.New("keygen").
 		Funcs(template.FuncMap{
 			"isChecked": func(checked bool) string {
@@ -184,15 +259,11 @@ func handleKeyGen(s *Service) http.HandlerFunc {
 	}
 
 	return func(w http.ResponseWriter, r *http.Request) {
-
 		f := keygenForm{Sub: true}
 		switch r.Method {
 		case "GET":
-			// do nothing
 		case "POST":
-
 			ok := f.parse(r)
-
 			if ok {
 				if f.isValid() {
 					key, err := s.generateKey(f.Key, f.Channel, f.access(), f.expires())
