+++ internal/security/cipher/xtea_test.go
@@ -133,42 +133,3 @@ func TestNewXtea(t *testing.T) {
 		}
 	}
 }
-
-/*
-func TestGenerateKey(t *testing.T) {
-	license, _ := ParseLicense("pLcaYvemMQOZR9o9sa5COWztxfAAAAAAAAAAAAAAAAI")
-	cipher, _ := license.Cipher()
-	masterKey, _ := cipher.DecryptKey([]byte("xEbaDPaICEwVhgdnl2rg_1DWi_MAg_3B"))
-
-	tests := []struct {
-		channel  string
-		channels []string
-		expected string
-		err      bool
-	}{
-		{channel: "article1", err: true},
-		{channel: "article1/", expected: "jhdrak0aHbbK6TbmyA391ndW3JwwgtNw", channels: []string{"article1/"}},
-		{channel: "article1/#/", expected: "jhdrak0aHbbRDL1NpzzN4HdW3JwwgtNw", channels: []string{"article1/", "article1/a/", "article1/a/b/c/", "article1/+/a/b/c/"}},
-	}
-
-	for _, tc := range tests {
-		key, err := cipher.GenerateKey(masterKey, tc.channel, AllowRead, time.Unix(0, 0), 1)
-		assert.Equal(t, tc.err, err != nil)
-		if tc.err {
-			continue
-		}
-
-		// Assert the key
-		assert.Equal(t, tc.expected, key)
-
-		// Attempt to parse the key
-		for _, c := range tc.channels {
-			channel := ParseChannel([]byte(tc.expected + "/" + c))
-			k, err := cipher.DecryptKey([]byte(key))
-			assert.NoError(t, err)
-			assert.True(t, k.ValidateChannel(channel))
-		}
-
-	}
-}
-*/
